refactor: extract skill system types
